# openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/agent-configurations/":{"get":{"tags":["agent-configurations"],"summary":"Get Configurations","operationId":"get_configurations_agent_configurations__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AgentConfigurationRead"},"type":"array","title":"Response Get Configurations Agent Configurations  Get"}}}}}},"post":{"tags":["agent-configurations"],"summary":"Create Configuration","operationId":"create_configuration_agent_configurations__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentConfigurationCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","format":"uuid","title":"Response Create Configuration Agent Configurations  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agent-configurations/{configuration_id}":{"get":{"tags":["agent-configurations"],"summary":"Get Configuration","operationId":"get_configuration_agent_configurations__configuration_id__get","parameters":[{"name":"configuration_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Configuration Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentConfigurationRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agent-configurations/upsert":{"put":{"tags":["agent-configurations"],"summary":"Upsert Configuration","operationId":"upsert_configuration_agent_configurations_upsert_put","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentConfigurationUpdate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentConfigurationRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agent-executions/":{"post":{"tags":["agent-executions"],"summary":"Create Execution","operationId":"create_execution_agent_executions__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentExecutionCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"string","format":"uuid","title":"Response Create Execution Agent Executions  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/agent-executions/{execution_id}":{"get":{"tags":["agent-executions"],"summary":"Get Execution","operationId":"get_execution_agent_executions__execution_id__get","parameters":[{"name":"execution_id","in":"path","required":true,"schema":{"type":"string","format":"uuid","title":"Execution Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgentExecutionRead"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AgentConfigurationCreate":{"properties":{"agent_type":{"type":"string","const":"search_agent","title":"Agent Type","description":"Agent type"},"data":{"additionalProperties":true,"type":"object","title":"Data","description":"Agent configuration data"}},"type":"object","required":["agent_type","data"],"title":"AgentConfigurationCreate"},"AgentConfigurationRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id","description":"Agent ID"},"agent_type":{"type":"string","const":"search_agent","title":"Agent Type","description":"Agent type"},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"Agent creation time"},"updated_at":{"type":"string","format":"date-time","title":"Updated At","description":"Agent update time"},"data":{"additionalProperties":true,"type":"object","title":"Data","description":"Agent configuration data"}},"type":"object","required":["id","agent_type","created_at","updated_at","data"],"title":"AgentConfigurationRead"},"AgentConfigurationUpdate":{"properties":{"id":{"type":"string","format":"uuid","title":"Id","description":"Agent ID"},"agent_type":{"type":"string","const":"search_agent","title":"Agent Type","description":"Agent type"},"data":{"additionalProperties":true,"type":"object","title":"Data","description":"Agent configuration data"}},"type":"object","required":["id","agent_type","data"],"title":"AgentConfigurationUpdate"},"AgentExecutionCreate":{"properties":{"id":{"type":"string","format":"uuid","title":"Id","description":"Agent ID"},"config_id":{"type":"string","format":"uuid","title":"Config Id","description":"Agent configuration ID"}},"type":"object","required":["id","config_id"],"title":"AgentExecutionCreate"},"AgentExecutionRead":{"properties":{"id":{"type":"string","format":"uuid","title":"Id","description":"Agent ID"},"config_id":{"type":"string","format":"uuid","title":"Config Id","description":"Agent configuration ID"},"executions":{"type":"integer","title":"Executions","description":"Agent executions"},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"Agent execution creation time"},"updated_at":{"type":"string","format":"date-time","title":"Updated At","description":"Agent execution update time"},"success_result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Success Result","description":"Success result"},"error_result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Error Result","description":"Error result"}},"type":"object","required":["id","config_id","executions","created_at","updated_at","success_result","error_result"],"title":"AgentExecutionRead"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}